refactor: support sub-command
support sub-command to add another sub-command (crank oracle feed)

fragmetric-snapshot orca-liquidity pool tokenA tokenB --rpc '' --output ''
-> fragmetric-snapshot snapshot orca-liquidity pool tokenA tokenB --rpc '' --output ''
